% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{american_date_format}
\alias{american_date_format}
\title{Format a date according to American convention}
\usage{
american_date_format(date)
}
\arguments{
\item{date}{A \link{Date} class vector, or an object coercible to one}
}
\value{
A character vector giving the American formatted version of the
dates in the object given as the \code{dates} argument
}
\description{
Puts a date in Month Date, Year format and removes any leading zeroes
}
