% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach.R
\name{attach_common_packages}
\alias{attach_common_packages}
\title{Attach packages commonly used in data cleaning and analysis}
\usage{
attach_common_packages(additional_packages = NULL, packages_to_omit = NULL)
}
\arguments{
\item{additional_packages}{A character vector giving the names of additional
packages to attach, or \code{NULL} (the default) if no additional packages
are needed.}

\item{packages_to_omit}{A character vector giving the names of packages
\emph{not} to attach, or \code{NULL} (the default) if no packages need be omitted.}
}
\value{
Invisibly returns a character vector giving the names of the
packages successfully attached
}
\description{
By default this function will attach the packages
\code{dplyr}, \code{tidyr}, \code{readr}, and \code{ggplot2}.
You can avoid attaching any of these packages via the \code{packages_to_omit}
parameter.
If any additional packages are requested via the \code{additional_packages}
parameter, those packages will be attached as well.
Information will be printed to the console listing which packages were
attached, which could not be attached, and what functions from the attached
packages mask other functions that were on the search path before calling
\code{attach_common_packages()}.
}
\examples{
\dontrun{
attach_common_packages()
}

}
