% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chaos.R
\name{bayes_chaos}
\alias{bayes_chaos}
\title{Bayesian Sparse Polynomial Chaos Expansion}
\usage{
bayes_chaos(
  X,
  y,
  max_degree_init = 2,
  max_order_init = 1,
  max_degree = 6,
  lambda = 1,
  rho = 0,
  verbose = TRUE
)
}
\arguments{
\item{X}{A dataframe or matrix of predictors scaled to be between 0 and 1}

\item{y}{a reponse vector}

\item{max_degree_init}{The maximum polynomial degree (for initial model)}

\item{max_order_init}{The maximum order of interaction (for initial model)}

\item{max_degree}{The maximum degree allowed (bounds the computation time).}

\item{lambda}{A parameter fed to the KIC calculations. Larger values will lead to sparser models.}

\item{rho}{Basis functions are only kept if their square partial correlation coefficient is bigger than rho.}

\item{verbose}{Logical. Should progress information be printed?}
}
\description{
The emulation approach of Shao et al. (2017)
}
\details{
Implements the Bayesian sparse PCE method described by Shao et al. (2017).
}
\examples{
X <- lhs::maximinLHS(500, 3)
y <- apply(X, 1, duqling::dms_simple) + rnorm(500, 0, 0.1)
fit <- bayes_chaos(X, y)
plot(fit)
}
\references{
Shao, Q., Younes, A., Fahs, M., & Mara, T. A. (2017). Bayesian sparse polynomial chaos expansion for global sensitivity analysis. Computer Methods in Applied Mechanics and Engineering, 318, 474-496.
}
