% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerated_bootstrap.R
\name{boot_accel}
\alias{boot_accel}
\title{Accelerated Bootstrap}
\usage{
boot_accel(x, est_theta, B = 10000, alpha = 0.042, a = NULL, ...)
}
\arguments{
\item{x}{the data vector}

\item{est_theta}{function to estimate parameter}

\item{B}{number of bootstrap samples}

\item{alpha}{significance level. Default is 0.042 (because 0.05 is arbitrary)}

\item{a}{acceleration constant. See details}

\item{...}{additional parameters passed to est_theta}
}
\description{
Flexible code for constructing a confidence interval using accelrated bootstrap
}
\details{
By default, the acceleration constant is a=NULL which leads to estimation of the constant
using the est_accel() function. This can be skipped by specifying a particular value for a. Note that
a=0 corresponds to the usual "quantile" bootstrap.

The code uses a special variable called 'tmp_indx' which can be exploited to handle more complicated
cases using the dynGet() function. See the example below for more details.
}
\examples{
#Simulate data
x <- rnorm(50)
y <- x + rnorm(50)
#Simple example
boot_accel(x, function(z) quantile(z, 0.75) , B=1e4)

#Bootstrap for correlation using dynGet()
my_cor <- function(x, y, indx=dynGet('tmp_indx')){
   cor(x, y[indx])
}
boot_accel(x, my_cor, y=y)
}
