% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/references.R
\name{categorize_references}
\alias{categorize_references}
\title{Determine what types of sources and which journals are cited in a bib file}
\usage{
categorize_references(filename)
}
\arguments{
\item{filename}{A character vector of length one giving the path of the bib
file you want examined}
}
\value{
Invisibly returns the contents of the bib file
}
\description{
This function reads in a bib file and outputs to the console which types of
sources are contained in it (@Article, @Book, etc), and how many of each are
in the bib file. It also outputs which journals the @Article references come
from, and how many times each journal is listed in an @Article reference in
the bib file.
}
