% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_distances.R
\name{cdf_dist_lp}
\alias{cdf_dist_lp}
\title{Lp Distance Between CDFs for Multivariate Samples}
\usage{
cdf_dist_lp(x, y, p = 1, n_pts = 30)
}
\arguments{
\item{x}{a vector of samples from P}

\item{y}{a vector of samples from Q}

\item{p}{norm order (p=Inf returns KS distance, p=1 returns Kantorovich distance)}

\item{n_pts}{number of evaluation points IN EACH DIMENSION. Run time is O(n_pts^ncol(x)).}
}
\description{
Lp Distance Between CDFs for Multivariate Samples
}
\details{
Approximates distance between CDFs. Note that the performance rapidly deteriorates for large ncol(x)
}
\examples{
x <- matrix(rnorm(300), ncol=3)
y <- matrix(rgamma(180, 3, 2) - 1.5, ncol=3)
cdf_dist_lp(x, y, p=1, n_pts=10)
cdf_dist_lp(x, y, p=2, n_pts=10)
cdf_dist_lp(x, y, p=Inf, n_pts=10)

}
