% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_research_project}
\alias{check_research_project}
\title{Check research project for best practices}
\usage{
check_research_project(path = getwd())
}
\arguments{
\item{path}{The path to the project's root directory; the default is the
current working directory}
}
\value{
Invisibly returns a list with the following elements:
\describe{
\item{unlisted_files}{
Files discovered in your project directory that are not described
in your README file
}
\item{current_R_version}{The value \code{\link[base]{R.version.string}}}
\item{unlisted_packages}{
A dataframe with the names and version numbers of
R packages detected in your research project's code that are either
not listed in your README file, or which are listed in your README
file in a way that does not reference the version of the package
currently installed on your machine.
}
\item{files_for_other_software}{
A character vector that includes
\itemize{
\item Paths to Python scripts if you do not mention in the
README file that Python was used in analysis, and
\item Paths to Stata do files if you do not mention in the
README file that Stata was used in analysis.
}
}
}
}
\description{
Check research project for best practices
}
