% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_functions.R
\name{corr_matrix}
\alias{corr_matrix}
\title{Correlation Matrix}
\usage{
corr_matrix(X, beta, corr = list(type = "exponential", power = 2))
}
\arguments{
\item{X}{nxd matrix of design points}

\item{beta}{correlation parameter. If 1/kappa is length scale then beta=log10(kappa).}

\item{corr}{list specifying the correlation function. See GPfit package for details.}
}
\description{
Genersates the power exponential or Matern correlation matrix for a set of n design points in d-dimensional space.
Default is Gaussian (power=2). See GPfit::corr_matrix for details.
}
\examples{
R <- corr_matrix(seq(0, 1, length.out=10), beta=log10(1.0))

}
