% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crps.R
\name{crps}
\alias{crps}
\title{Estimate CRPS(F, y) given samples from F}
\usage{
crps(y, x, w = 0)
}
\arguments{
\item{y}{the true scalar value}

\item{x}{a vector of samples from F}

\item{w}{parameter (between 0 and 1). When w=0, the estimator is unbiased ("fair" and "PWM" from Zamo & Naveau 2017). When w=1, the estimator has lower variance ("INT" and "NRG" estimator from Zamo & Naveau 2017).}
}
\description{
Returns the an estimator between the INT and PWM estimators of Zamo
}
\examples{
F1 <- function(x) pnorm(x)
I1 <- function(x, y) as.numeric(x >= y)
Q <- function(x, y) (F1(x) - I1(x, y))^2
y <- 0.1
x <- rnorm(100)
integrate(Q, lower=-10, upper=10, y=y)$value
crps(y, x, w=0) #default w
crps(y, x, w=0.5)
crps(y, x, w=1)
}
\references{
Zamo, M., & Naveau, P. (2018). Estimation of the continuous ranked probability score with limited information and applications to ensemble weather forecasts. Mathematical Geosciences, 50(2), 209-234.
}
