% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint_CR.R
\name{d_mahal}
\alias{d_mahal}
\title{Mahalanobis Distance}
\usage{
d_mahal(x, y = NULL, S = diag(rep(1, length(x))))
}
\arguments{
\item{x}{a vector}

\item{y}{an optional vector. The zero vector by default}

\item{S}{a precision matrix of dimension equal to length(x)}
}
\description{
Computes the mahalanobis distance between two vectors given a precision matrix S
}
\examples{
x <- rnorm(5); y <- rnorm(5); S <- GPfit::corr_matrix((1:5)/6, beta=0.2)
d_mahal(x, y, S)
}
