% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{density01}
\alias{density01}
\title{Kernel Density Estimate with Beta Kernels}
\usage{
density01(x, v = 100, n = 512)
}
\arguments{
\item{x}{the data from which the estimate is to be computed.}

\item{v}{a smoothness parameter (smaller is smoother)}

\item{n}{the number of points in (0, 1) on which the density is evaluated}
}
\value{
a list with components x and y
}
\description{
A kernel density estimate which will always integrate to 1 on the interval (0, 1)
}
\details{
KDE with beta kernels centered at each point in x. v is the "sample size" parameter of the beta.
}
\examples{
x = rbeta(150, 3, 2)
hist(x, freq=F, breaks=30, xlim=c(0,1))
lines(density01(x, v=50), lwd=2, col='firebrick')
lines(density01(x, v=100), lwd=2, col='orange')
lines(density01(x, v=1000), lwd=2, col='dodgerblue')
}
