% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MP.R
\name{diagnose_ppc}
\alias{diagnose_ppc}
\title{Diagnostic Plot for Probability of Prior Coherency}
\usage{
diagnose_ppc(M, V, p, MC = 10000, control = NULL, ...)
}
\arguments{
\item{M}{a vector of posterior M values (mean of a parameter set)}

\item{V}{a vector of posterior V values (variance of a parameter set)}

\item{p}{dimension of the parameter set}

\item{MC}{number of Monte Carlo samples.}

\item{control}{a list of control parameters for plotting. See details.}

\item{...}{additional parameters passed to plot() function.}
}
\description{
Diagnostic plot for detecting overfitting using probability of prior coherency
}
\details{
control is a named list consisting of any of the following.

levels - the desired contour levels for plotting (default c(0.5, 0.9, 0.95, 0.99))

h - step-size for plotting contours (default 0.005)

tol - tolerance for plotting contours (default 1e-7)

prob - a probability bound for setting xlim and ylim (default 0.9999)

fill_col - the color of the contours regions (default 'dodgerblue')

point_col - the color of the points (default 'orange')

cex - 0.2 the size of the points representing PPC posterior samples (default 0.5)

leg_cex - the size of the legend (default 1.4)
}
\examples{
X <- rMP(100, 7, 1, 3)
}
