% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn_mix_EM.R
\name{dmvn_mix}
\alias{dmvn_mix}
\title{Density of mvn_mix object}
\usage{
dmvn_mix(x, obj, log = FALSE)
}
\arguments{
\item{x}{vector of locations to evaluate density}

\item{obj}{an element of the list generated by mvn_mix()

Chi, Eric C., and Kenneth Lange. "Stable estimation of a covariance matrix guided by nuclear norm penalties." Computational statistics & data analysis 80 (2014): 117-128.}
}
\description{
Density of mvn_mix object
}
\examples{
x <- rgamma(100, 3, 1.2)
fit <- mvn_mix(x, K=1:3)
obj <- fit[[3]]
xx <- rmvn_mix(1000, obj)
hist(xx)
dmvnorm(c(2, 3, 4), obj)
}
