% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECP.R
\name{ECP}
\alias{ECP}
\title{Emulation of the Conditional Posterior (General)}
\usage{
ECP(
  lpost,
  L = 30,
  init = NULL,
  MCMC = list(pars = NULL, iterations = 5000, burn_in = 1000, thin = 1, prop_sigma =
    NULL, adapt_par = c(100, 20, 0.5, 0.75)),
  f_alpha = function(x) x[1],
  estimate_psi = function(x) c(mean(x), sd(x)),
  gamma_generate = NULL,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{lpost}{The log posterior. First argument must be a vector of parameters for which
full Bayesian inference is desired. Second argument is a vector of nuisance parameters
for which modularization is desired. Second argument must be named "gamma".}

\item{L}{the budget.}

\item{init}{a set of initial values for alpha. Can also be a function which generates starting values
based on gamma}

\item{MCMC}{a named list of arguments for the Metro_Hastings function.}

\item{f_alpha}{an optional function for transforming the physical parameters. Helpful for reducing
to a single dimension.}

\item{estimate_psi}{function whose argument is the result of f_alpha(al), where al is a
posterior sample from the conditional posterior. Should return a vector of r parameter estimates.
Defaults to univariate normal.}

\item{gamma_generate}{can be either (i) an L x q matrix of nuisance parameters, (ii) a function
(with no arguments) that generates a single instance of gamma (length q) from the prior or (iii)
a named list with two components (mu and sigma) which are q-length vectors of prior mean/sd's.
Please ensure that class(gamma_generate) is either 'matrix', 'function' or 'list'.}

\item{...}{additional arguments passed to lpost}
}
\description{
This function models the conditional posterior distribution of alpha conditional on a
set of nuisance parameters gamma. Draws from the modularization posterior can be obtained
by pairing this function with ECP_sample. This function is designed to give maximal flexibility.
Consider using ECP_norm for a more straightforward implementation and ECP_multi for multiple alpha
}
\examples{
set.seed(42)
a <- 0; g <- 0.1
n1 <- 10; n2 <- 90
y <- rnorm(n1+n2, a + c(rep(0, n1), rep(g, n2)), 1)
lpost <- function(a, gamma, y, n1, n2){
  res <- sum(dnorm(y, a + c(rep(0, n1), rep(gamma, n2)), 1, log=TRUE))
  res <- res + dnorm(a, 0, 1, log=TRUE) + dnorm(gamma, 0, 1, log=TRUE)
  return(res)
}
ecp <- ECP(lpost, L=30, init=0, gamma_generate=list(mu=0, sigma=1), y=y, n1=n1, n2=n2)
hist(ECP_sample(ecp, M=1000, gamma_generate=list(mu=0, sigma=1)), breaks=30)
ECP_profile_plot(ecp)
}
