% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECP.R
\name{ECP_profile_plot}
\alias{ECP_profile_plot}
\title{Profile Plot for ECP Algorithm}
\usage{
ECP_profile_plot(
  ecp,
  bounds = c(0, 1, 0.01),
  fix = NULL,
  quantiles = c(0.95, 0.5, 0.05),
  ...
)
}
\arguments{
\item{ecp}{an object generated from the ecp function (requires normality)}

\item{bounds}{defines the bounds of gamma (from, to, by).
Should be a q x 3 matrix when there are multiple gamma.}

\item{fix}{when q > 1, what values should the remaining modularization parameters be fixed to?}

\item{quantiles}{the quantiles of the conditional distribution to plot}

\item{...}{additional arguments passed to plot() and lines()}
}
\description{
Plots quantiles of the posterior for alpha conditional on gamma.
}
\examples{
set.seed(42)
a <- 0; g <- 0.1
n1 <- 10; n2 <- 90
y <- rnorm(n1+n2, a + c(rep(0, n1), rep(g, n2)), 1)
lpost <- function(a, gamma, y, n1, n2){
  res <- sum(dnorm(y, a + c(rep(0, n1), rep(gamma, n2)), 1, log=TRUE))
  res <- res + dnorm(a, 0, 1, log=TRUE) + dnorm(gamma, 0, 1, log=TRUE)
  return(res)
}
ecp <- ECP(lpost, L=30, init=0, gamma_generate=list(mu=0, sigma=1), y=y, n1=n1, n2=n2)
hist(ECP_sample(ecp, M=1000, gamma_generate=list(mu=0, sigma=1)), breaks=30)
ECP_profile_plot(ecp)
}
