% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECP.R
\name{ECP_sample}
\alias{ECP_sample}
\title{Sampling from the Modularization Posterior with ECP}
\usage{
ECP_sample(ecp, M = 1, gamma_generate = NULL, alpha_generate = NULL)
}
\arguments{
\item{ecp}{An object returned from the ECP() function.}

\item{M}{number of samples requested}

\item{gamma_generate}{can be either (i) an L x q matrix of nuisance parameters, (ii) a function
(with no arguments) that generates a single instance of gamma (length q) from the prior or (iii)
a named list with two components (mu and sigma) which are q-length vectors of prior mean/sd's.
Please ensure that class(gamma_generate) is either 'matrix', 'function' or 'list'.}

\item{alpha_generate}{a function to generate random samples of alpha conditional on psi. Should
correspond to the estimate_psi function used in ECP() function. Takes one argument: psi and returns
a single alpha draw. Defaults to univariate normal.}
}
\description{
This function produces draws from the Modularization Posterior of alpha (with respect to gamma).
}
\examples{
set.seed(42)
a <- 0; g <- 0.1
n1 <- 10; n2 <- 90
y <- rnorm(n1+n2, a + c(rep(0, n1), rep(g, n2)), 1)
lpost <- function(a, gamma, y, n1, n2){
  res <- sum(dnorm(y, a + c(rep(0, n1), rep(gamma, n2)), 1, log=TRUE))
  res <- res + dnorm(a, 0, 1, log=TRUE) + dnorm(gamma, 0, 1, log=TRUE)
  return(res)
}
ecp <- ECP(lpost, L=30, init=0, gamma_generate=list(mu=0, sigma=1), y=y, n1=n1, n2=n2)
hist(ECP_sample(ecp, M=1000, gamma_generate=list(mu=0, sigma=1)), breaks=30)
ECP_profile_plot(ecp)
}
