% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egamma.R
\name{egamma}
\alias{egamma}
\alias{degamma}
\alias{pegamma}
\alias{qegamma}
\alias{regamma}
\alias{megamma}
\title{Extended Gamma Distribution

The EGamma(alpha, beta, kappa) is best described by its CDF which has the form F(x) = wF_1(x) + (1-w)F_2(x)
where w = beta^-alpha/(beta^-alpha - (beta + kappa)^-alpha) > 1. F_1 is the CDF of a gamma(alpha, beta) distribution
and F_2 is the CDF of a gamma(alpha, beta + kappa) distribution.}
\usage{
degamma(x, alpha, beta, kappa, log = FALSE)

pegamma(q, alpha, beta, kappa, log = FALSE, lower.tail = TRUE)

qegamma(p, alpha, beta, kappa, ...)

regamma(n, alpha, beta, kappa, ...)

megamma(k = 1, alpha, beta, kappa, ...)
}
\arguments{
\item{x}{a vector of length p}

\item{alpha}{shape of the gamma distribution}

\item{beta}{rate of the gamma distribution}

\item{kappa}{extension parameter (positive)}

\item{log}{logical. Should density/probability be returned on a log scale?}

\item{q}{a vector of quantiles}

\item{lower.tail}{logical; if TRUE (default) probabilities are P(X <= x), otherwise P(X > x).}

\item{p}{a vector of probabilities}

\item{...}{optional additional arguments passed to \code{optim()}}

\item{n}{number of observations. If \code{length(n) > 1}, the length is taken to be the number required.}

\item{k}{the moment desired}
}
\description{
Extended Gamma Distribution

The EGamma(alpha, beta, kappa) is best described by its CDF which has the form F(x) = wF_1(x) + (1-w)F_2(x)
where w = beta^-alpha/(beta^-alpha - (beta + kappa)^-alpha) > 1. F_1 is the CDF of a gamma(alpha, beta) distribution
and F_2 is the CDF of a gamma(alpha, beta + kappa) distribution.
}
\details{
\code{degamma} returns the density, \code{pegamma} returns the distribution function, \code{qegamma} returns the quantile function, \code{regamma} generates random samples (using inverse transform) and \code{megamma} returns the moments of the distribution.
}
\examples{
n <- 100 #Number of observations
X <- regamma(n, 3, 1, 2)
hist(X, breaks=20, freq=FALSE)
curve(degamma(x, 3, 1, 2), add=TRUE)
}
