% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerated_bootstrap.R
\name{est_accel}
\alias{est_accel}
\title{Estimation of the "Accelration Constant"}
\usage{
est_accel(x, est_theta, ...)
}
\arguments{
\item{x}{the data vector}

\item{est_theta}{function to estimate parameter}
}
\description{
Estimates the accleration constant using a simple jackknife estimator (Efron 1987).
}
\examples{
est_accle(rnorm(30), function(z) quantile(z, 0.75))
}
