% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_functions.R
\name{fast_det}
\alias{fast_det}
\title{Fast Matrix Algebra for BMC}
\usage{
fast_det(FM, phi = NULL, tau = NULL, log = TRUE)
}
\arguments{
\item{FM}{an object created by the fast_process() function}

\item{phi}{current value of phi (can be omitted if phi is stored in FM object)}

\item{tau}{current value of tau (can be omitted if tau is stored in FM object)}

\item{log}{logical. Should log-determinant be returned? Default is TRUE.}
}
\value{
Returns the (log) determinant of the covariance matrix.
}
\description{
Function for obtaining the determinant in linear time - after pre-processing - for covariance matrices of suitable structure
}
\examples{
n <- 100
R <- corr_matrix(seq(0,1,length.out=n), beta=0.2)
FM <- fast_process(R, method='approx', control=list(tol=1e-9))
#Compute determinant (approximately)
fast_det(FM, phi=1.5, tau=0.2)
#Compute inverse (approximately)
fast_inv(FM, phi=1.5, tau=0.2)
}
