% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_functions.R
\name{fast_inv}
\alias{fast_inv}
\title{Fast Matrix Algebra for BMC}
\usage{
fast_inv(FM, ...)
}
\arguments{
\item{FM}{an object of class "FM_fast", "FM_approx" or "FM_ts"}

\item{...}{appropriate arguments for the corresponding class.}
}
\value{
Returns (possibly an approximation of) the inverse of the covariance matrix
}
\description{
See fast_inv.class for details.
}
