% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastMatrixBMC.R
\name{fast_process}
\alias{fast_process}
\title{Fast Matrix Algebra for BMC}
\usage{
fast_process(R, phi = NULL, tau = NULL, method = "fast", control = list())
}
\arguments{
\item{R}{fixed correlation matrix}

\item{phi}{Optional. Indicates that phi is fixed throughout the analysis.}

\item{tau}{Optional. Indicates that tau is fixed throughout the analysis. Required when method == 'ts'}

\item{method}{Must be 'fast', 'approx', or 'ts'.}

\item{control}{a list of control parameters. Will be ignored when method == 'fast'.
Method 'approx' needs control$tol (1e-9) recomended, and method 'ts' needs an expansion value control$a and an expansion order control$P}
}
\value{
returns an object of class "FM_<method>" containing the information needed for fast computation of matrix inverse and determinant
}
\description{
Functions for obtaining matrix inversions and determinants quickly in the context of BMC.
Processes matrices of the form Sigma = phi x R + tau x I
}
\examples{
n <- 100
R <- GPfit::corr_matrix(seq(0,1,length.out=n), beta=0.2)
FM <- fast_process(R, method='approx', control=list(tol=1e-9))
#Compute determinant (approximately)
fast_det(FM, phi=1.5, tau=0.2)
#Compute inverse (approximately)
fast_inv(FM, phi=1.5, tau=0.2)

}
