% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MP.R
\name{get_constMP}
\alias{get_constMP}
\title{Normalizing Constant for the Moment Penalization Prior}
\usage{
get_constMP(p, w1 = 1, w2 = 1, MC = 1e+07)
}
\arguments{
\item{p}{the dimension of the vector having an MP prior}

\item{w1}{normalized penalty associated with second moment. Default is 1}

\item{MC}{number of MC iterations. Default is 1e7, but larger values are recommended when speed is not important.}
}
\value{
returns the normalizing constant of the MP prior
}
\description{
Returns the normalizing constant for the Moment Penalization Prior with parameters p, w1 and w2
}
\examples{
get_constMP(7, 1, 3, MC=1e9)
}
