% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{ggsummary}
\alias{ggsummary}
\title{A summary function to make things readily available for ggplot}
\usage{
ggsummary(
  data = NULL,
  measurevar,
  groupvars = NULL,
  na.rm = FALSE,
  conf.interval = 0.95,
  .drop = TRUE
)
}
\arguments{
\item{data}{the data frame}

\item{measurevar}{the response variable (in quotes)}

\item{groupvars}{a vector of grouping variables (in quotes)}

\item{na.rm}{should na values be removed}

\item{conf.interval}{coverage probability desired for t-intervals}
}
\value{
a data.frame of mean,sd,ci,n grouped by groupvars
}
\description{
Requires plyr pacakge
}
\details{
A function for easy handling of data frame to be used for ggplot2
}
\examples{
set.seed(111)
df <- data.frame(foo = sample(c("a", "b", "c"), 100, replace=TRUE),
                 fie = rpois(100, 0.5))
df$y <- rnorm(100, df$fie + 1, 0.1)*(df$foo == "a") +
        rnorm(100, df$fie + 2, 0.1)*(df$foo == "b") +
        rnorm(100, df$fie + 3, 0.1)*(df$foo == "c")

summ <- ggsummary(df, measurevar = "y", groupvars = c("foo", "fie"))
#ggplot(summ, aes(x=fie, y=mean, fill=foo)) +
#     geom_bar(position=position_dodge2(padding=0.1), stat="identity") +
#     geom_errorbar(aes(ymin=mean-ci, ymax=mean+ci),
#   width=.05,
#     position=position_dodge(.9))

}
