% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint_CR.R
\name{inside_CR}
\alias{inside_CR}
\title{Is Inside Credible Region}
\usage{
inside_CR(q, CR)
}
\arguments{
\item{q}{a query point}

\item{CR}{the vertices of the convex hull returned by joint_CR (with two_dim = TRUE)}
}
\value{
Returns +1, 0, or -1 for inside JCR, on JCR or outside JCR respectively.
}
\description{
Checks to see whether a 2-D query point is inside, on or outside the convex hull
representing an empirical credible region
}
\details{
Only implemented (currently) for 2 dimensions.
}
