% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint_CR.R
\name{joint_CR}
\alias{joint_CR}
\title{Joint Confidence Regions using Mahalanobis Distance}
\usage{
joint_CR(X, alpha = 0.042, two_dim = TRUE)
}
\arguments{
\item{X}{a matrix - each row represents a sample}

\item{alpha}{the confidence level is 1-alpha. Defaults to 0.042 (because 0.05 is so arbitrary)}

\item{two_dim}{logical. If TRUE, then the convex hull surrounding the interior points in the first two dimensions is returned.
If FALSE, then we return the entire set of interior points (for arbitrary dimension).}
}
\value{
the samples laying on the convex hull which contains (1-alpha)100\% of the samples.
}
\description{
Finds the minimal convex hull, with respect to Mahalanobis distance, which contains (1-alpha)x100\% of the points.
}
\details{
Note that a set of N-dimensional points is always returned. When two_dim=TRUE,
this set of points represents the points on the convex hull of the JCR - but only in the first two dimensions.
This is a limitation of the built-in chull() function. Future versions may extend the functionality.
For now, one can operate in higher dimensions by setting two_dim = FALSE to return the full set of
(1-alpha)*N points which are in the interior of the JCR.
}
\examples{
x <- rnorm(1000); y <- x + rnorm(1000, 0, .5)
X <- cbind(x, y)
CR <- joint_CR(X)
plot(X)
polygon(CR, border='blue', lwd=2)

#A point inside of JCR
inside_CR(c(0,0), CR)
#A point on the convex hull of JCR
lambda <- 0.3
inside_CR(lambda*CR[3,] + (1-lambda)*CR[4,], CR)
#A point outside of JCR
inside_CR(c(-3, 7), CR)
}
