% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_distances.R
\name{kantorovich_dist}
\alias{kantorovich_dist}
\title{Kantorovich Distance Between CDFs for Multivariate Samples}
\usage{
kantorovich_dist(x, y, n_pts = 30)
}
\arguments{
\item{x}{a vector of samples from P}

\item{y}{a vector of samples from Q}

\item{n_pts}{number of evaluation points IN EACH DIMENSION. Run time is O(n_pts^ncol(x)).}
}
\description{
Kantorovich Distance Between CDFs for Multivariate Samples
}
\details{
Multivariate KS distance
}
\examples{
x <- matrix(rnorm(300), ncol=3)
y <- matrix(rgamma(180, 3, 2) - 1.5, ncol=3)
kantorovich_dist(x, y)

}
