% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kld.R
\name{kld}
\alias{kld}
\title{KL Divergence based on samples from P and Q}
\usage{
kld(x, y, type = 1, lambda = 1/2, n_mc = 1)
}
\arguments{
\item{x}{a vector of samples from P}

\item{y}{a vector of samples from Q}

\item{type}{See details}

\item{lambda}{parameter (between 0 and 1) for the Population Stability Index (ignored unless type = 4)}

\item{n_mc}{number of monte carlo samples used when type = 3 or 4}
}
\description{
Returns the KLD estimator of Perez-Cruz (2008)
}
\details{
Set type = 1 for D(P||Q), type = 2 for symmetrized "Jeffreys Divergence", type = 3 for symmetric "Jensen-Shannon Distance", type = 4 for Population Stability Index (with parameter lambda between 0 and 1). When type = 3 or 4, n_mc must be set for approximation. When type = 4, lambda must be specified. lambda = 1/2 is equivalent to type = 3.
}
\examples{
x <- rnorm(100)
y <- rgamma(200, 3, 2) - 1.5
kld(x, y)

}
