% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leapGP.R
\name{leapGP_synch}
\alias{leapGP_synch}
\title{Localized Ensemble of Approximate Gaussian Processes}
\usage{
leapGP_synch(leapGP, rho = 0.95)
}
\arguments{
\item{leapGP}{an object of class "leapGP"}

\item{rho}{parameter controlling time-accuracy tradeoff (default = 0.95)}
}
\value{
an object of class "leapGP" AND of class "slapGP"
}
\description{
This function synchronizes an object of class "leapGP", so that is also an object of "slapGP"
}
\examples{
Xnew <- matrix(runif(100), nrow=50, ncol=2)
X <- matrix(runif(100), nrow=50, ncol=2)
Y <- apply(X, 1, prod)
preds1 <- leapGP(Xnew, X, Y)

#Or equivalently
leap <- leapGP_build(X, Y)
preds2 <- rep(NA, 100)
for(i in 1:100){
   preds2[i] <- leapGP_predict(leap, Xnew[i,])
}

#Or used with slapGP
leap <- leapGP_synch(leap, rho=0.95)
hubs <- leap$hubs
preds3 <- rep(NA, 100)
for(i in 1:100){
   emulator <- slapGP(Xnew[i,], X, Y, hubs=hubs)
   preds3[m] <- emulator$pred
   hubs <- emulator$hubs
}
}
