% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_common_packages}
\alias{load_common_packages}
\title{Load packages commonly used in data cleaning and analysis}
\usage{
load_common_packages(additional_packages = NULL, packages_to_omit = NULL)
}
\arguments{
\item{additional_packages}{A character vector giving the names of additional
packages to load, or \code{NULL} (the default) if no additional packages are
needed.}

\item{packages_to_omit}{A character vector giving the names of packages
\emph{not} to load, or \code{NULL} (the default) if no packages need be omitted.}
}
\value{
Invisibly returns a character vector giving the names of the
packages loaded
}
\description{
By default this function will load the packages
\href{dplyr::dplyr}{dplyr},
\href{tidyr::tidyr}{tidyr},
\href{readr::readr}{readr},
and \href{ggplot2::ggplot2}{ggplot2}.
You can avoid loading any of these packages via the \code{packages_to_omit}
parameter.
If any additional packages are requested via the \code{additional_packages}
parameter, those packages will be loaded as well.
If any of the packages the function attempts to load are not installed,
the function will issue a warning that the unavailable package(s) could not
be loaded.
}
\examples{
\dontrun{
load_common_packages()
}

}
