% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeGraphs.R
\name{makeGraphs}
\alias{makeGraphs}
\title{Make your graphs function}
\usage{
makeGraphs(usr_network_file, usr_network_name, usr_save_loc)
}
\arguments{
\item{usr_network_file, usr_network_name, usr_save_loc}{Characters referring to the location and file name of user's network file, name of network, and location to save output.}
}
\value{
rdata file
}
\description{
This function allows user to upload network file and create graph function compatible with Quack.
}
\examples{
makeGraphc("data/InWeb3.txt", "InWeb3", "data/")
}
