% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maximal_coupling.R
\name{maximal_coupling}
\alias{maximal_coupling}
\title{Maximal Couplings}
\usage{
maximal_coupling(n, d1, d2, g1 = NULL, g2 = NULL)
}
\arguments{
\item{n}{number of requested samples}

\item{d1}{density of distribution 1}

\item{d2}{density of distribution 2}

\item{g1}{generator for distribution 1}

\item{g2}{generator for distribution 2}
}
\value{
An nx2 matrix of draws where the ith column has (marginal) distribution i. The Pr(X\link{,1} == X\link{,2}) is maximized.
}
\description{
A function to sample from the maximal coupling of two distributions (using Johnson (1998))
}
\details{
Uses the gamma-coupling algorithm of Johnson 1998
}
\examples{
d1 <- function(x) dnorm(x)
d2 <- function(x) dunif(x, -3, 3)
g1 <- function()  rnorm(1)
g2 <- function()  runif(1, -3, 3)
maximal_coupling(1000, d1, d2, g1, g2)
}
