% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn_mix_EM.R
\name{mvn_mix}
\alias{mvn_mix}
\title{EM Algorithm to fit Finite Mixtures of Multivariate Normal Distributions}
\usage{
mvn_mix(
  X,
  K = 1:3,
  kappa = 0.75,
  epsilon = 1e-05,
  max_iter = 2000,
  verbose = FALSE,
  print_iter = 10,
  crit = NULL
)
}
\arguments{
\item{X}{data, a nxp matrix}

\item{K}{a vector (or scalar for a single fit) of}

\item{kappa}{Regularization parameter. Each Sigma is shrank towards Sx at iteration t with weight 1/t^kappa}

\item{epsilon}{stopping criterion for EM algorithm}

\item{max_iter}{maximum number of iterations of EM algorithm}

\item{verbose}{should information on status be printed}

\item{print_iter}{how often should information being printed? Ignored when verbose=FALSE.}

\item{crit}{a subset of c("aic", "bic", "hqic") corresponding to Aikike's, Bayesian and Hanaan-Quinn information criteria.}
}
\value{
a list with components corresponding to the fit using the number of components specified by \code{K}
}
\description{
Returns a list of mixture distribution for various number of mixture components
}
\details{
A regularized approach based loosely on Chi & Lange (2014).

Chi, Eric C., and Kenneth Lange. "Stable estimation of a covariance matrix guided by nuclear norm penalties." Computational statistics & data analysis 80 (2014): 117-128.
}
\examples{
n1 <- 300
mu1 <- c(0, 0)
sigma1 <- matrix(c(1, 0.5, 0.5, 1), nrow=2, byrow=TRUE)
X1 <- mvtnorm::rmvnorm(n1, mu1, sigma1)

n2 <- 500
mu2 <- c(1, 1.5)
sigma2 <- matrix(c(1, -0.7, -0.7, 1), nrow=2, byrow=TRUE)
X2 <- mvtnorm::rmvnorm(n2, mu2, sigma2)

X <- rbind(X1, X2)
plot(X)

#Fit mixture models for 1 to 3 compoenents
obj <- mvn_mix(X, K=1:3, epsilon=1e-3, verbose=TRUE, crit=c("aic", "bic", "hqic"))

plot(1:3, obj$aic, type='l', col="orange")
lines(1:3, obj$bic, col="dodgerblue")
lines(1:3, obj$hqic, col="firebrick")
legend("bottomleft", c("aic", "bic", "hqic"), lwd=2, col=c("orange", "dodgerblue", "firebrick"))
}
