% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MP.R
\name{ppc}
\alias{ppc}
\title{Probability of Prior Coherency}
\usage{
ppc(M, V, p, type = 1, MC = 10000)
}
\arguments{
\item{M}{a vector of posterior M values (mean of a parameter set)}

\item{V}{a vector of posterior V values (variance of a parameter set)}

\item{p}{dimension of the parameter set}

\item{type}{indicates what should be returned. See details}

\item{MC}{number of Monte Carlo samples.}
}
\description{
Computes the probability of prior coherency using Monte Carlo
}
\details{
type = 1 leads to PPC calculation based on posterior mean of M and V (as described in paper)

type = 2 is the posterior mean of PPC

type = 3 returns posterior samples of PPC

type = 4 returns additional information (used for plotting)
}
\examples{
gamma <- cbind(rnorm(1000, 1.2, 0.2),
               rnorm(1000, -0.4, 0.3),
               rnorm(1000, -0.6, 0.15))
M <- apply(gamma, 1, mean)
V <- apply(gamma, 1, var)
ppc(M, V, 3)
diagnose_ppc(M, V, p, control=list(levels=c(0.5, 0.9, 0.99), fill_col='purple'))
}
