% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chaos.R
\name{predict.bayes_chaos}
\alias{predict.bayes_chaos}
\title{Predict Method for class bayes_chaos}
\usage{
\method{predict}{bayes_chaos}(object, newdata = NULL, samples = FALSE)
}
\arguments{
\item{object}{An object returned by the \code{bayes_chaos()} function.}

\item{newdata}{A dataframe of the same dimension as the training data.}

\item{samples}{How many posterior samples should be taken at each test point? If 0 or FALSE, then the MAP estimate is returned.}
}
\description{
See \code{bayes_chaos()} for details.
}
\details{
Predict function for bayes_chaos object.
}
\examples{
X <- lhs::maximinLHS(500, 3)
y <- apply(X, 1, duqling::dms_simple) + rnorm(500, 0, 0.1)
fit <- bayes_chaos(X, y)
pred <- predict(fit, X)
}
\references{
Shao, Q., Younes, A., Fahs, M., & Mara, T. A. (2017). Bayesian sparse polynomial chaos expansion for global sensitivity analysis. Computer Methods in Applied Mechanics and Engineering, 318, 474-496.
}
