% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFok.R
\name{predict.rfok}
\alias{predict.rfok}
\title{Predict method for rfok}
\usage{
\method{predict}{rfok}(object, newdata, conf = 0.95, samples = NULL, ...)
}
\arguments{
\item{object}{Object returned by \code{rfok()}}

\item{newdata}{a data frame or matrix of new data}

\item{conf}{a vector of desired confidence intervals}

\item{samples}{Number of samples from the predictive distribution.}
}
\value{
An object with class "rfok"
}
\description{
A conformal RF based on Johansson et al. (2014). Rather than using a traditional calibration set, conformal intervals are constructed using out-of-bag samples for each tree.
}
\examples{

X <- matrix(runif(150), nrow=50, ncol=3)
y <- apply(X, 1, duqling::ishigami)
fit <- rfok(X, y)

Xnew <- matrix(runif(150), nrow=50, ncol=3)
predict(fit, Xnew)
}
