% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{presentation}
\alias{presentation}
\title{Custom formatted Beamer presentation}
\usage{
presentation(...)
}
\arguments{
\item{...}{Arguments to be passed to \code{\link[bookdown:html_document2]{bookdown::beamer_presentation2()}}}
}
\value{
An R Markdown output format object
}
\description{
Custom formatted Beamer presentation
}
