% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list-packages.R
\name{print.RequiredPackages}
\alias{print.RequiredPackages}
\title{Print a RequiredPackages object}
\usage{
\method{print}{RequiredPackages}(x, ...)
}
\arguments{
\item{x}{An object of class \link[=list_required_packages]{RequiredPackages}}

\item{...}{Arguments passed to or from other methods (currently unused)}
}
\value{
Invisibly returns x
}
\description{
Prints the information from the RequiredPackages object in a format that may
be used as a Markdown unordered list.
}
