% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quack-package.R
\docType{package}
\name{quack-package}
\alias{quack}
\alias{quack-package}
\title{quack: Research Project Setup and Management Tools}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Don't be a 'quackademic'! Manage your research projects in an organized and consistent way. This package provides an R Project template to setup research project directories in a consistent way conveniently.
}
\author{
\strong{Maintainer}: JBrandon Duck-Mayr \email{j.duckmayr@gmail.com} (\href{https://orcid.org/0000-0002-2231-1294}{ORCID})

}
\keyword{internal}
