% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quackPrediction.R
\name{quackPrediction}
\alias{quackPrediction}
\title{Quack prediction function}
\usage{
quackPrediction(usr_network_name, usr_seed_file, trained_network_file,
  quack_gene_model, usr_output_loc, usr_pathway_file)
}
\arguments{
\item{usr_network_name}{Characters of user's network name, used in make graphs and train model functions.}

\item{usr_seed_file}{Characters of path and name of user's seed file containing list of genes.}

\item{trained_network_file}{Characters of path and name to network file from make graphs function.}

\item{quack_gene_model}{Characters of path and name to quack gene model created from train model function.}

\item{usr_output_loc}{Characters of location to save output file.}

\item{usr_pathway_file}{Characters of name of pathway files, which can be found in data/ or user uploaded.}
}
\value{
text file
}
\description{
Using the trained model from user uploaded network, using seed genes, predict genes that should be in a network.
}
\examples{
quackPrediction("InWeb3", "data/SandersGenes.txt", "data/InWeb3.RData", "QuackGeneModel_Quackv1.3-InWeb3-General853StringentPathways.RData", "data/", "data/stringent853Pathways.RData")
}
