% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFok.R
\name{rfok}
\alias{rfok}
\title{Conformal Random Forest with Out-of-bag KNN}
\usage{
rfok(X, y, k = 5, beta = sd(y)/30, ...)
}
\arguments{
\item{X}{a data frame or a matrix of predictors}

\item{y}{a response vector}

\item{k}{Number of nearest neighbors to use for out-of-bag error estimaters}

\item{beta}{Sensitivity parameter}
}
\value{
An object with class "rfok"
}
\description{
A conformal RF based on Johansson et al. (2014). Rather than using a traditional calibration set, conformal intervals are constructed using out-of-bag samples for each tree.
}
\examples{

X <- matrix(runif(150), nrow=50, ncol=3)
y <- apply(X, 1, duqling::ishigami)
fit <- rfok(X, y)

Xnew <- matrix(runif(150), nrow=50, ncol=3)
predict(fit, Xnew)
}
\references{
Johansson, U., Boström, H., Löfström, T., & Linusson, H. (2014). Regression conformal prediction with random forests. Machine learning, 97, 155-176.
}
