% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MP.R
\name{rMP}
\alias{rMP}
\title{Random Sampling from the Moment Penalization Prior}
\usage{
rMP(n, p, w1, w2, acceptance = FALSE)
}
\arguments{
\item{n}{number of samples requested}

\item{p}{dimension of random vector}

\item{w1}{normalized penalty associated with second moment. Default is 1}

\item{acceptance}{logical. If TRUE, a list is returned with an attribute giving the acceptance rate.}

\item{...}{additional parameters passed to get_constMP (if norm=TRUE)}
}
\value{
returns the density of the MP(w1, w2) prior
}
\description{
Uses an optimized rejection sampling framework to sample from the MP prior
}
\examples{
X <- rMP(100, 7, 1, 3)
}
