% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MP.R
\name{rZreg}
\alias{rZreg}
\title{Random Sampling from Z-Regularization Prior}
\usage{
rZreg(n, p, sigma_R = 0)
}
\arguments{
\item{n}{number of samples requested}

\item{p}{dimension of random vector}

\item{sigma_R}{relaxation parameter (default 0)}
}
\value{
returns the density of the Z regularization prior
}
\description{
Draws samples from Z-regularization prior. Alternative to MP prior with w1=w2=Inf (when sigma_R = 0).
}
\examples{
X <- rZreg(100, 7)
}
