% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{smartLHS}
\alias{smartLHS}
\title{Customized "smart" (robust) LHS function}
\usage{
smartLHS(n, k, maximin_n = 1000, ...)
}
\arguments{
\item{n}{the number of rows or samples}

\item{k}{the number of columns or parameters/variables}

\item{maximin_n}{the number of rows or samples in the original maximin design}

\item{...}{additional arguments passed to randomLHS (n > 50k) or maximinLHS (n <= 50k)}
}
\value{
a latin hypercube design
}
\description{
A default lhs function to use (to avoid accidental crashing of maximinLHS)
}
\details{
Tries to use a maxmin LHS design when possible (n <= maximin_n = 1000).
If n is greater than 100,000, we simply return a random LHS.
In between these cases, we build an initial maximin design of size maximin_n and then augment this design with augmentLHS.
}
\examples{
x <- myLHS(1000, 8)
y <- apply(x, 1, quack::ff_borehole)
}
