% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{submission}
\alias{submission}
\title{PDF document formatted to meet APSR, AJPS, and JOP submission standards}
\usage{
submission(...)
}
\arguments{
\item{...}{Arguments to be passed to \code{\link[bookdown:html_document2]{bookdown::pdf_document2()}}}
}
\value{
An R Markdown output format object
}
\description{
PDF document formatted to meet APSR, AJPS, and JOP submission standards
}
