% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_functions.R
\name{test_functions}
\alias{test_functions}
\alias{ff_borehole}
\alias{ff_lim}
\alias{ff_gl2008}
\alias{ff_ripples}
\alias{ff_synth_3level}
\alias{get_input_length}
\title{Some general testing functions for UQ}
\usage{
ff_borehole(x, scale01 = TRUE)

ff_lim(x)

ff_gl2008(x, scale01 = TRUE)

ff_ripples(x)

ff_synth_3level(x)

get_input_length(function_name)
}
\arguments{
\item{x}{a vector in inputs. type get_input_length("function_name") to get the minimum requirement for length(x)}

\item{scale01}{default TRUE. Expects inputs on \link{0,1} space. When FALSE, inputs are expected in there natural ranges}
}
\description{
Some general testing functions for UQ
}
\details{
List of functions and details
\enumerate{
\item borehole
\item \href{https://www.sfu.ca/~ssurjano/limetal02pol.html}{lim}
\item (gramacy_lee)\link{https://www.sfu.ca/~ssurjano/grlee08.html}
\item synth_3level
}
}
\examples{
n <- 10
x <- matrix(runif(3*n), nrow=n)
y <- apply(x, 1, f_lim)
}
