% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainModel.R
\name{trainModel}
\alias{trainModel}
\title{Train model function}
\usage{
trainModel(usr_modeling_file, usr_pathway_file, usr_output_loc)
}
\arguments{
\item{usr_modeling_file, usr_pathway_file, usr_output_loc}{Characters referring to the location and file name of user's network modeling file, path and name of pathway files, which can be found in data/ or user uploaded, and location to save output.}
}
\value{
rdata file, csv and txt file
}
\description{
Using user uploaded network and known pathways, train model using Quack algorithm.
}
\examples{
trainModel("data/InWeb3.RData", "data/stringent835Pathways.RData", "data/")
}
