% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FastMatrixBMC2.R
\name{fast_inv.FM_fast}
\alias{fast_inv.FM_fast}
\title{Fast Matrix Algebra for BMC}
\usage{
fast_inv.FM_fast(FM, phi = NULL, tau = NULL)
}
\arguments{
\item{FM}{an object of class "FM_fast"}

\item{phi}{current value of phi (can be omitted if phi is stored in FM object)}

\item{tau}{current value of tau (can be omitted if tau is stored in FM object)}
}
\value{
Returns the inverse of the covariance matrix
}
\description{
Function for computing the inverse of a BMC Covariance matrix. Leads to speedup of roughly 2 when matrix dimension is moderate or large.
}
